refactor: Share the `MediaSession` fingerprints
