--- patches/src/main/kotlin/app/morphe/patches/music/interaction/scrobbling/Fingerprints.kt
@@ -1,31 +1,0 @@
-/*
- * Copyright 2026 Morphe.
- * https://github.com/MorpheApp/morphe-patches
- *
- * See the included NOTICE file for GPLv3 Section 7 terms that apply to this code.
- */
-
-package app.morphe.patches.music.interaction.scrobbling
-
-import app.morphe.patcher.Fingerprint
-import app.morphe.patcher.methodCall
-
-internal object MediaSessionSetPlaybackStateFingerprint : Fingerprint(
-    filters = listOf(
-        methodCall(
-            definingClass = "Landroid/media/session/MediaSession;",
-            name = "setPlaybackState",
-            parameters = listOf("Landroid/media/session/PlaybackState;")
-        )
-    )
-)
-
-internal object MediaSessionSetMetadataFingerprint : Fingerprint(
-    filters = listOf(
-        methodCall(
-            definingClass = "Landroid/media/session/MediaSession;",
-            name = "setMetadata",
-            parameters = listOf("Landroid/media/MediaMetadata;")
-        )
-    )
-)
