--- patches/src/main/kotlin/app/morphe/patches/music/layout/lyrics/Fingerprints.kt
@@ -1,33 +1,0 @@
-/*
- * Copyright 2026 Morphe.
- * https://github.com/MorpheApp/morphe-patches/pull/2269
- *
- * See the included NOTICE file for GPLv3 Section 7 terms that apply to this code.
- */
-
-package app.morphe.patches.music.layout.lyrics
-
-import app.morphe.patcher.Fingerprint
-import app.morphe.patcher.methodCall
-
-// Own instances, not the scrobbling ones: a fingerprint caches its match, and that
-// patch inserts an instruction at the matched index, shifting the index used here.
-internal object LyricsMediaSessionSetMetadataFingerprint : Fingerprint(
-    filters = listOf(
-        methodCall(
-            definingClass = "Landroid/media/session/MediaSession;",
-            name = "setMetadata",
-            parameters = listOf("Landroid/media/MediaMetadata;")
-        )
-    )
-)
-
-internal object LyricsMediaSessionSetPlaybackStateFingerprint : Fingerprint(
-    filters = listOf(
-        methodCall(
-            definingClass = "Landroid/media/session/MediaSession;",
-            name = "setPlaybackState",
-            parameters = listOf("Landroid/media/session/PlaybackState;")
-        )
-    )
-)
