+++ patches/src/main/kotlin/app/morphe/patches/music/layout/lyrics/LyricsPatch.kt
@@ -7,8 +7,6 @@
 
 package app.morphe.patches.music.layout.lyrics
 
-import app.morphe.patcher.extensions.InstructionExtensions.addInstruction
-import app.morphe.patcher.extensions.InstructionExtensions.getInstruction
 import app.morphe.patcher.patch.bytecodePatch
 import app.morphe.patches.all.misc.resources.addResourcesPatch
 import app.morphe.patches.music.misc.extension.sharedExtensionPatch
@@ -18,10 +16,12 @@ import app.morphe.patches.music.misc.settings.settingsPatch
 import app.morphe.patches.music.shared.Constants.COMPATIBILITY_YOUTUBE_MUSIC
 import app.morphe.patches.music.video.information.musicVideoInformationPatch
 import app.morphe.patches.shared.misc.litho.filter.addLithoFilter
+import app.morphe.patches.shared.misc.media.MediaSessionSetMetadataFingerprint
+import app.morphe.patches.shared.misc.media.MediaSessionSetPlaybackStateFingerprint
+import app.morphe.patches.shared.misc.media.hookMediaSessionArgument
 import app.morphe.patches.shared.misc.settings.preference.ListPreference
 import app.morphe.patches.shared.misc.settings.preference.NonInteractivePreference
 import app.morphe.patches.shared.misc.settings.preference.SwitchPreference
-import com.android.tools.smali.dexlib2.iface.instruction.FiveRegisterInstruction
 
 private const val EXTENSION_CLASS = "Lapp/morphe/extension/music/patches/lyrics/LyricsPatch;"
 
@@ -71,28 +71,12 @@ val lyricsPatch = bytecodePatch(
         // lyrics component is the earliest signal that the opened panel is the lyrics one.
         addLithoFilter(LYRICS_PANEL_FILTER)
 
-        LyricsMediaSessionSetMetadataFingerprint.let {
-            it.method.apply {
-                val index = it.instructionMatches.first().index
-                val register = getInstruction<FiveRegisterInstruction>(index).registerD
-                addInstruction(
-                    index,
-                    "invoke-static { v$register }, $EXTENSION_CLASS->" +
-                            "onSetMetadata(Landroid/media/MediaMetadata;)V"
-                )
-            }
-        }
+        MediaSessionSetMetadataFingerprint.hookMediaSessionArgument(
+            "$EXTENSION_CLASS->onSetMetadata(Landroid/media/MediaMetadata;)V"
+        )
 
-        LyricsMediaSessionSetPlaybackStateFingerprint.let {
-            it.method.apply {
-                val index = it.instructionMatches.first().index
-                val register = getInstruction<FiveRegisterInstruction>(index).registerD
-                addInstruction(
-                    index,
-                    "invoke-static { v$register }, $EXTENSION_CLASS->" +
-                            "onSetPlaybackState(Landroid/media/session/PlaybackState;)V"
-                )
-            }
-        }
+        MediaSessionSetPlaybackStateFingerprint.hookMediaSessionArgument(
+            "$EXTENSION_CLASS->onSetPlaybackState(Landroid/media/session/PlaybackState;)V"
+        )
     }
 }
