--- patches/src/main/kotlin/app/morphe/patches/youtube/misc/medianotification/Fingerprints.kt
@@ -1,14 +1,0 @@
-package app.morphe.patches.youtube.misc.medianotification
-
-import app.morphe.patcher.Fingerprint
-import app.morphe.patcher.methodCall
-
-internal object MediaSessionSetPlaybackStateFingerprint : Fingerprint(
-    filters = listOf(
-        methodCall(
-            definingClass = "Landroid/media/session/MediaSession;",
-            name = "setPlaybackState",
-            parameters = listOf("Landroid/media/session/PlaybackState;")
-        )
-    )
-)
